Fix: report.js to ts
